% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceclust.R
\name{CECclassifNewData}
\alias{CECclassifNewData}
\title{Assignment of new individuals to existing clusters}
\usage{
CECclassifNewData(Zpred, params, idColToPred = c())
}
\arguments{
\item{Zpred}{data.frame or matrix of new observations
(same structure as \code{Z} used for training, except for columns to predict).}

\item{params}{list. Model parameters returned by \code{\link[=CECclassif]{CECclassif()}}.}

\item{idColToPred}{integer. Indices of columns ignored for assignment,
not included in Zpred.}
}
\value{
An integer vector of cluster indices
(length \code{nrow(Zpred)}).
}
\description{
Given a \code{params} object learned by \code{\link[=CECclassif]{CECclassif()}},
assigns each row of \code{Zpred} to the most likely cluster.
}
\examples{
# continuation of the CECclassif() example
nnew <- 10
# Two Gaussian clusters on 2 variables
x1 <- rnorm(nnew, 0, 1); x2 <- rnorm(nnew, 0, 1)
x3 <- rnorm(nnew, 3, 1); x4 <- rnorm(nnew, 3, 1)
X  <- rbind(cbind(x1, x2), cbind(x3, x4))
# A categorical variable correlated with the cluster
lab <- factor(rep(c("A","B"), each = nnew))

Znew <- data.frame(X1 = X[,1], X2 = X[,2])
pred_clusters <- CECclassifNewData(Znew, fit$params, idColToPred = 3)
pred_clusters

}
\seealso{
\code{\link[=CECclassif]{CECclassif()}}, \code{\link[=CECpredict]{CECpredict()}}
}
