\name{CEClust-package}
\alias{CEClust}
\alias{CEClust-package}
\title{CEClust: Clustering par Entropie Composite}
\description{
Le package \pkg{CEClust} implémente un algorithme de clustering pour données mixtes (numériques et catégorielles) basé sur l'entropie composite.
}
\details{
Les fonctions principales sont \code{\link{CECclassif}}, \code{\link{CECclassifNewData}} et \code{\link{CECpredict}}.
}
\author{Thierry Dumont <thierry.dumont@example.com>}
\keyword{package}
